+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Improve progress notification when tool is running.
+
 ## 0.58.0
 
 - Bump MCP java sdk to 0.13.1